重构任务栏歌词界面结构
