+++ src/renderer-taskbar-lyric/App.vue
@@ -8,303 +8,58 @@
     @contextmenu.prevent="handleContextMenu"
     @pointerenter="handlePointerEnter"
     @pointerleave="handlePointerLeave"
+    @dragstart.prevent
   >
     <div v-if="state.showCover" class="cover">
-      <img v-if="state.albumCoverUrl" :src="state.albumCoverUrl" alt="album cover">
+      <img v-if="state.albumCoverUrl" :src="state.albumCoverUrl" alt="album cover" draggable="false">
       <div v-else class="cover-fallback">LX</div>
     </div>
     <div class="content">
-      <template v-if="showActionButtons">
-        <div class="action-buttons" @dblclick.stop>
-          <button type="button" class="action-button" title="上一首" aria-label="上一首" @pointerdown.stop @click.stop="handleActionClick('prev')">
-            <svg viewBox="0 0 20 20" aria-hidden="true">
-              <path d="M5.2 4.3a1 1 0 0 1 1 1v9.4a1 1 0 1 1-2 0V5.3a1 1 0 0 1 1-1Zm9.28.58a1 1 0 0 1-.1 1.62L9.6 10l4.78 3.5a1 1 0 0 1-1.18 1.62l-5.86-4.28a1 1 0 0 1 0-1.62l5.86-4.28a1 1 0 0 1 1.28.08Z" fill="currentColor" />
-            </svg>
-          </button>
-          <button
-            type="button"
-            class="action-button action-button-primary"
-            :title="state.isPlaying ? '暂停' : '播放'"
-            :aria-label="state.isPlaying ? '暂停' : '播放'"
-            @pointerdown.stop
-            @click.stop="handleActionClick(state.isPlaying ? 'pause' : 'play')"
-          >
-            <svg v-if="state.isPlaying" viewBox="0 0 20 20" aria-hidden="true">
-              <path d="M6 4.5A1.5 1.5 0 0 1 7.5 6v8A1.5 1.5 0 0 1 6 15.5 1.5 1.5 0 0 1 4.5 14V6A1.5 1.5 0 0 1 6 4.5Zm8 0A1.5 1.5 0 0 1 15.5 6v8A1.5 1.5 0 0 1 14 15.5 1.5 1.5 0 0 1 12.5 14V6A1.5 1.5 0 0 1 14 4.5Z" fill="currentColor" />
-            </svg>
-            <svg v-else viewBox="0 0 20 20" aria-hidden="true">
-              <path d="M6.24 4.86c0-.93 1.02-1.5 1.81-1.01l7.12 4.43a2.02 2.02 0 0 1 0 3.44l-7.12 4.43c-.8.49-1.8-.08-1.8-1.01V4.86Z" fill="currentColor" />
-            </svg>
-          </button>
-          <button type="button" class="action-button" title="下一首" aria-label="下一首" @pointerdown.stop @click.stop="handleActionClick('next')">
-            <svg viewBox="0 0 20 20" aria-hidden="true">
-              <path d="M14.8 4.3a1 1 0 0 1 1 1v9.4a1 1 0 1 1-2 0V5.3a1 1 0 0 1 1-1Zm-9.28.58a1 1 0 0 1 1.28-.08l5.86 4.28a1 1 0 0 1 0 1.62L6.8 15.12a1 1 0 1 1-1.18-1.62L10.4 10 5.62 6.5a1 1 0 0 1-.1-1.62Z" fill="currentColor" />
-            </svg>
-          </button>
-        </div>
-      </template>
+      <TaskbarLyricActionButtons v-if="showActionButtons" :is-playing="state.isPlaying" @action="handleActionClick" />
       <template v-else>
-        <div v-if="state.showSongInfo" ref="songInfoRef" class="song-info">
-          <div
-            v-if="shouldScrollSongInfo"
-            class="song-info-track"
-            :style="songInfoTrackStyle"
-          >
-            <span class="song-info-text">{{ displaySongInfoText }}</span>
-            <span class="song-info-gap" aria-hidden="true"></span>
-            <span class="song-info-text" aria-hidden="true">{{ displaySongInfoText }}</span>
-          </div>
-          <template v-else>
-            <span class="title">{{ primarySongInfoText }}</span>
-            <span v-if="secondarySongInfoText" class="separator">-</span>
-            <span v-if="secondarySongInfoText" class="artist">{{ secondarySongInfoText }}</span>
-          </template>
-          <span ref="songInfoMeasureRef" class="song-info-measure">{{ displaySongInfoText }}</span>
-        </div>
-        <div v-if="state.showCurrentLine" ref="lyricLineRef" class="lyric-line">
-          <div
-            v-if="shouldScrollLyric"
-            class="lyric-line-track"
-            :style="lyricLineTrackStyle"
-          >
-            <span class="lyric-line-text">{{ displayLyricText }}</span>
-            <span class="lyric-line-gap" aria-hidden="true"></span>
-            <span class="lyric-line-text" aria-hidden="true">{{ displayLyricText }}</span>
-          </div>
-          <span v-else class="lyric-line-text">{{ displayLyricText }}</span>
-          <span ref="lyricMeasureRef" class="lyric-line-measure">{{ displayLyricText }}</span>
-        </div>
+        <TaskbarLyricSongInfo
+          v-if="state.showSongInfo"
+          :primary-text="primarySongInfoText"
+          :secondary-text="secondarySongInfoText"
+          :font-size="state.songInfoFontSize"
+        />
+        <TaskbarLyricLyricLine
+          v-if="state.showCurrentLine"
+          :text="displayLyricText"
+          :font-size="state.lyricFontSize"
+        />
       </template>
     </div>
   </div>
 </template>
 
 <script setup lang="ts">
+import { computed, onBeforeUnmount } from 'vue'
+import TaskbarLyricActionButtons from './components/TaskbarLyricActionButtons.vue'
+import TaskbarLyricLyricLine from './components/TaskbarLyricLyricLine.vue'
+import TaskbarLyricSongInfo from './components/TaskbarLyricSongInfo.vue'
+import { useTaskbarLyricShellStyle } from './composables/useTaskbarLyricShellStyle'
+import { useTaskbarLyricWindowDrag } from './composables/useTaskbarLyricWindowDrag'
 import { state } from './store/state'
-import { computed, nextTick, onBeforeUnmount, onMounted, ref, watch } from 'vue'
-import { requestTaskbarLyricMenu, requestTaskbarLyricShowMainInterface, sendTaskbarLyricControl, sendTaskbarLyricDragEnd, sendTaskbarLyricDragMove } from './utils/ipc'
-
-interface RGB {
-  r: number
-  g: number
-  b: number
-}
+import { sendTaskbarLyricControl } from './utils/ipc'
 
 const lyricState = state as LX.TaskbarLyric.State
-const isDragging = ref(false)
-const isHovering = ref(false)
-const songInfoRef = ref<HTMLElement | null>(null)
-const songInfoMeasureRef = ref<HTMLElement | null>(null)
-const lyricLineRef = ref<HTMLElement | null>(null)
-const lyricMeasureRef = ref<HTMLElement | null>(null)
-const shouldScrollSongInfo = ref(false)
-const songInfoScrollDistance = ref(0)
-const shouldScrollLyric = ref(false)
-const lyricScrollDistance = ref(0)
-let pointerId: number | null = null
-let startScreenX = 0
-let startOffsetX = 0
-let lyricResizeObserver: ResizeObserver | null = null
+const { shellStyle } = useTaskbarLyricShellStyle()
+const {
+  isDragging,
+  isHovering,
+  handlePointerDown,
+  handlePointerEnter,
+  handlePointerLeave,
+  handleContextMenu,
+  handleDoubleClick,
+  stopDragging,
+} = useTaskbarLyricWindowDrag()
 
-const showActionButtons = computed(() => {
-  return isHovering.value && !isDragging.value
-})
+const showActionButtons = computed(() => isHovering.value && !isDragging.value)
 const primarySongInfoText = computed(() => lyricState.swapTitleAndArtist && lyricState.artist ? lyricState.artist : lyricState.title)
 const secondarySongInfoText = computed(() => lyricState.swapTitleAndArtist ? lyricState.title : lyricState.artist)
-const displaySongInfoText = computed(() => {
-  return secondarySongInfoText.value ? `${primarySongInfoText.value} - ${secondarySongInfoText.value}` : primarySongInfoText.value
-})
 const displayLyricText = computed(() => lyricState.lyricLine || lyricState.artist)
-const songInfoTrackStyle = computed(() => {
-  const distance = Math.max(songInfoScrollDistance.value, 0)
-  const gap = 24
-  const duration = Math.max(10, distance / 26)
-  return {
-    '--taskbar-song-info-scroll-distance': `${distance + gap}px`,
-    '--taskbar-song-info-scroll-duration': `${duration.toFixed(2)}s`,
-  }
-})
-const lyricLineTrackStyle = computed(() => {
-  const distance = Math.max(lyricScrollDistance.value, 0)
-  const gap = 24
-  const duration = Math.max(8, distance / 28)
-  return {
-    '--taskbar-lyric-line-scroll-distance': `${distance + gap}px`,
-    '--taskbar-lyric-line-scroll-duration': `${duration.toFixed(2)}s`,
-  }
-})
-
-const clamp = (value: number, min: number, max: number) => Math.min(Math.max(value, min), max)
-
-const parseRgb = (color: string | null | undefined): RGB | null => {
-  if (!color) return null
-  const value = color.trim()
-  if (!value) return null
-
-  const hex = value.replace(/^#/, '')
-  if (/^[\da-f]{3}$/i.test(hex)) {
-    return {
-      r: parseInt(hex[0] + hex[0], 16),
-      g: parseInt(hex[1] + hex[1], 16),
-      b: parseInt(hex[2] + hex[2], 16),
-    }
-  }
-  if (/^[\da-f]{6}$/i.test(hex)) {
-    return {
-      r: parseInt(hex.slice(0, 2), 16),
-      g: parseInt(hex.slice(2, 4), 16),
-      b: parseInt(hex.slice(4, 6), 16),
-    }
-  }
-  const match = value.match(/rgba?\(([\d.]+)[, ]+([\d.]+)[, ]+([\d.]+)/i)
-  if (!match) return null
-  return {
-    r: clamp(Math.round(Number(match[1])), 0, 255),
-    g: clamp(Math.round(Number(match[2])), 0, 255),
-    b: clamp(Math.round(Number(match[3])), 0, 255),
-  }
-}
-
-const toRgbString = (color: RGB) => `rgb(${color.r}, ${color.g}, ${color.b})`
-const withAlpha = (color: RGB, alpha: number) => `rgba(${color.r}, ${color.g}, ${color.b}, ${clamp(alpha, 0, 1)})`
-
-const mix = (colorA: RGB, colorB: RGB, weight: number): RGB => {
-  const ratio = clamp(weight, 0, 1)
-  const remain = 1 - ratio
-  return {
-    r: Math.round(colorA.r * remain + colorB.r * ratio),
-    g: Math.round(colorA.g * remain + colorB.g * ratio),
-    b: Math.round(colorA.b * remain + colorB.b * ratio),
-  }
-}
-
-const getLuminance = ({ r, g, b }: RGB) => {
-  const normalize = (channel: number) => {
-    const value = channel / 255
-    return value <= 0.03928 ? value / 12.92 : ((value + 0.055) / 1.055) ** 2.4
-  }
-  const red = normalize(r)
-  const green = normalize(g)
-  const blue = normalize(b)
-  return red * 0.2126 + green * 0.7152 + blue * 0.0722
-}
-
-const getContrastRatio = (foreground: RGB, background: RGB) => {
-  const [lighter, darker] = [getLuminance(foreground), getLuminance(background)].sort((a, b) => b - a)
-  return (lighter + 0.05) / (darker + 0.05)
-}
-
-const getReadableTextColor = (background: RGB) => {
-  const lightText = { r: 248, g: 250, b: 252 }
-  const darkText = { r: 15, g: 23, b: 42 }
-  return getContrastRatio(lightText, background) >= getContrastRatio(darkText, background)
-    ? lightText
-    : darkText
-}
-
-const shellStyle = computed(() => {
-  const backgroundOpacity = clamp((lyricState.backgroundOpacity ?? 72) / 100, 0, 1)
-  const themeColor = parseRgb(lyricState.themeColor) ?? { r: 77, g: 175, b: 124 }
-  const isLightTheme = getLuminance(themeColor) > 0.58
-  const isCustomSongInfoFontMode = lyricState.songInfoFontColorMode === 'custom'
-  const isCustomLyricFontMode = lyricState.lyricFontColorMode === 'custom'
-
-  const backgroundBase = lyricState.backgroundColorMode === 'custom'
-    ? parseRgb(lyricState.backgroundColor) ?? themeColor
-    : isLightTheme
-      ? mix(themeColor, { r: 255, g: 255, b: 255 }, 0.82)
-      : mix(themeColor, { r: 15, g: 23, b: 42 }, 0.72)
-
-  const backgroundStrong = lyricState.backgroundColorMode === 'custom'
-    ? backgroundBase
-    : isLightTheme
-      ? mix(themeColor, { r: 255, g: 255, b: 255 }, 0.72)
-      : mix(themeColor, { r: 30, g: 41, b: 59 }, 0.6)
-
-  const defaultLyricText = getReadableTextColor(backgroundBase)
-  const defaultSongInfoText = mix(defaultLyricText, backgroundBase, 0.34)
-  const lyricText = isCustomLyricFontMode
-    ? parseRgb(lyricState.lyricFontColor) ?? defaultLyricText
-    : defaultLyricText
-  const songInfoText = isCustomSongInfoFontMode
-    ? parseRgb(lyricState.songInfoFontColor) ?? defaultSongInfoText
-    : defaultSongInfoText
-
-  const borderColor = mix(lyricText, backgroundBase, 0.76)
-  const borderOpacity = backgroundOpacity * 0.26
-
-  return {
-    '--taskbar-lyric-bg': withAlpha(backgroundBase, backgroundOpacity),
-    '--taskbar-lyric-bg-strong': withAlpha(backgroundStrong, backgroundOpacity),
-    '--taskbar-lyric-border': withAlpha(borderColor, borderOpacity),
-    '--taskbar-lyric-text': toRgbString(lyricText),
-    '--taskbar-lyric-text-secondary': toRgbString(songInfoText),
-    '--taskbar-lyric-song-info-font-size': `${clamp(lyricState.songInfoFontSize ?? 11, 9, 18)}px`,
-    '--taskbar-lyric-line-font-size': `${clamp(lyricState.lyricFontSize ?? 12, 10, 22)}px`,
-  }
-})
-
-const updateSongInfoScrollState = () => {
-  const containerWidth = songInfoRef.value?.clientWidth ?? 0
-  const contentWidth = songInfoMeasureRef.value?.scrollWidth ?? 0
-  const overflowWidth = Math.max(contentWidth - containerWidth, 0)
-  shouldScrollSongInfo.value = overflowWidth > 6
-  songInfoScrollDistance.value = overflowWidth
-}
-
-const updateLyricScrollState = () => {
-  const containerWidth = lyricLineRef.value?.clientWidth ?? 0
-  const contentWidth = lyricMeasureRef.value?.scrollWidth ?? 0
-  const overflowWidth = Math.max(contentWidth - containerWidth, 0)
-  shouldScrollLyric.value = overflowWidth > 6
-  lyricScrollDistance.value = overflowWidth
-}
-
-const handlePointerMove = (event: PointerEvent) => {
-  if (!isDragging.value || event.pointerId !== pointerId) return
-  const offsetX = startOffsetX + (event.screenX - startScreenX)
-  sendTaskbarLyricDragMove(offsetX)
-}
-
-const stopDragging = (event?: PointerEvent) => {
-  if (!isDragging.value) return
-  if (event && pointerId != null && event.pointerId !== pointerId) return
-  isDragging.value = false
-  pointerId = null
-  sendTaskbarLyricDragEnd()
-  window.removeEventListener('pointermove', handlePointerMove)
-  window.removeEventListener('pointerup', stopDragging)
-  window.removeEventListener('pointercancel', stopDragging)
-}
-
-const handlePointerEnter = () => {
-  isHovering.value = true
-}
-
-const handlePointerLeave = () => {
-  isHovering.value = false
-}
-
-const handlePointerDown = (event: PointerEvent) => {
-  if (event.button !== 0) return
-  isDragging.value = true
-  pointerId = event.pointerId
-  startScreenX = event.screenX
-  startOffsetX = state.offsetX
-  window.addEventListener('pointermove', handlePointerMove)
-  window.addEventListener('pointerup', stopDragging)
-  window.addEventListener('pointercancel', stopDragging)
-}
-
-const handleContextMenu = () => {
-  stopDragging()
-  requestTaskbarLyricMenu()
-}
-
-const handleDoubleClick = () => {
-  stopDragging()
-  requestTaskbarLyricShowMainInterface()
-}
 
 const handleActionClick = (action: 'prev' | 'next' | 'play' | 'pause') => {
   stopDragging()
@@ -312,55 +67,7 @@ const handleActionClick = (action: 'prev' | 'next' | 'play' | 'pause') => {
 }
 
 onBeforeUnmount(() => {
-  lyricResizeObserver?.disconnect()
-  lyricResizeObserver = null
   stopDragging()
-})
-
-onMounted(() => {
-  lyricResizeObserver = new ResizeObserver(() => {
-    updateSongInfoScrollState()
-    updateLyricScrollState()
-  })
-  if (songInfoRef.value) lyricResizeObserver.observe(songInfoRef.value)
-  if (songInfoMeasureRef.value) lyricResizeObserver.observe(songInfoMeasureRef.value)
-  if (lyricLineRef.value) lyricResizeObserver.observe(lyricLineRef.value)
-  if (lyricMeasureRef.value) lyricResizeObserver.observe(lyricMeasureRef.value)
-  void nextTick(() => {
-    updateSongInfoScrollState()
-    updateLyricScrollState()
-  })
-})
-
-watch(displaySongInfoText, () => {
-  void nextTick(() => {
-    updateSongInfoScrollState()
-  })
-})
-
-watch(displayLyricText, () => {
-  void nextTick(() => {
-    updateLyricScrollState()
-  })
-})
-
-watch(() => lyricState.songInfoFontSize, () => {
-  void nextTick(() => {
-    updateSongInfoScrollState()
-  })
-})
-
-watch(() => lyricState.lyricFontSize, () => {
-  void nextTick(() => {
-    updateLyricScrollState()
-  })
-})
-
-watch(showActionButtons, () => {
-  void nextTick(() => {
-    updateSongInfoScrollState()
-    updateLyricScrollState()
-  })
 })
 </script>
 
@@ -422,6 +129,8 @@ body {
     width: 100%;
     height: 100%;
     object-fit: cover;
+    user-select: none;
+    -webkit-user-drag: none;
   }
 }
 
@@ -476,9 +185,9 @@ body {
   }
 
   svg {
+    display: block;
     width: 15px;
     height: 15px;
-    display: block;
   }
 }
 
@@ -492,12 +201,12 @@ body {
   align-items: baseline;
   gap: 5px;
   min-width: 0;
+  overflow: hidden;
   color: var(--taskbar-lyric-text-secondary);
   font-size: var(--taskbar-lyric-song-info-font-size, 11px);
   line-height: 1.1;
-  opacity: 0.82;
-  overflow: hidden;
   white-space: nowrap;
+  opacity: 0.82;
 }
 
 .song-info-track {
@@ -547,11 +256,11 @@ body {
 .lyric-line {
   position: relative;
   margin: 0;
+  overflow: hidden;
   color: var(--taskbar-lyric-text);
   font-size: var(--taskbar-lyric-line-font-size, 12px);
   line-height: 1.1;
   font-weight: 500;
-  overflow: hidden;
   white-space: nowrap;
 }
